Move main function to core
This just seems cleaner if another __main__ uses this function
