+++ redbot/core/__init__.py
@@ -1,7 +1,8 @@
 import pkg_resources
 
-from .config import Config
-
-__all__ = ["Config", "__version__"]
-
 __version__ = version = pkg_resources.require("Red-DiscordBot")[0].version
+
+from .config import Config
+from .main import main
+
+__all__ = ["Config", "main", "__version__"]
